backend/requirements.txt
fastapi
uvicorn
python-dotenv
apify-client
